+++ test/storage.test.js
@@ -11,6 +11,7 @@
  */
 
 /* eslint-env mocha */
+/* eslint-disable max-classes-per-file */
 
 'use strict';
 
@@ -23,13 +24,84 @@ const { Response } = require('@adobe/helix-universal');
 const { AWSStorage } = proxyquire('../src/storage.js', {
   '@aws-sdk/client-s3': {
     S3Client: class {
-      constructor({ region }) {
-        this._region = region;
+      // eslint-disable-next-line class-methods-use-this
+      send(command) {
+        return command.run(this._storage);
       }
 
+      set storage(s) {
+        this._storage = s;
+      }
+    },
+    HeadBucketCommand: class {
+      constructor({ Bucket }) {
+        this._bucket = Bucket;
+      }
+
+      run(storage) {
+        if (storage.get(this._bucket)) {
+          return { $metadata: { httpStatusCode: 200 } };
+        }
+        const e = new Error();
+        e.$metadata = { httpStatusCode: 404 };
+        throw e;
+      }
+    },
+    CreateBucketCommand: class {
+      constructor({ Bucket }) {
+        this._bucket = Bucket;
+      }
+
+      run(storage) {
+        if (storage.get(this._bucket)) {
+          const e = new Error();
+          e.$metadata = { httpStatusCode: 409 };
+          throw e;
+        }
+        storage.set(this._bucket, []);
+        return { $metadata: { httpStatusCode: 200 } };
+      }
+    },
+    PutObjectCommand: class {
+      constructor({ Key, Bucket }) {
+        this._key = Key;
+        this._bucket = Bucket;
+      }
+
+      run(storage) {
+        const objs = storage.get(this._bucket);
+        if (!objs) {
+          const e = new Error();
+          e.$metadata = { httpStatusCode: 404 };
+          throw e;
+        }
+        objs.push(this._key);
+      }
+    },
+    GetBucketTaggingCommand: class {
+      constructor({ Bucket }) {
+        this._bucket = Bucket;
+      }
+
+      run(storage) {
+        if (storage.get(this._bucket)) {
+          return { TagSet: [] };
+        }
+        const e = new Error();
+        e.$metadata = { httpStatusCode: 404 };
+        throw e;
+      }
+    },
+    PutBucketTaggingCommand: class {
       // eslint-disable-next-line class-methods-use-this
-      async send(command) {
-        return command.input;
+      run() {
+        /* do nothing */
+      }
+    },
+    PutPublicAccessBlockCommand: class {
+      // eslint-disable-next-line class-methods-use-this
+      run() {
+        /* do nothing */
       }
     },
   },
@@ -38,23 +110,60 @@ const { AWSStorage } = proxyquire('../src/storage.js', {
 describe('Storage Tests', () => {
   it('constructor throws if required parameters are missing', async () => {
     assert.throws(() => new AWSStorage({}), /required/);
-    assert.throws(() => new AWSStorage({
-      AWS_S3_REGION: 'foo',
-    }), /required/);
-    assert.throws(() => new AWSStorage({
-      AWS_S3_REGION: 'foo',
-      AWS_S3_ACCESS_KEY_ID: 'bar',
-    }), /required/);
   });
-  it('actual', async () => {
+  it('constructor succeeds if required parameters are there', async () => {
+    assert.doesNotThrow(() => new AWSStorage({
+      mount: { url: 'mymount' },
+    }));
+  });
+  it('store item to non existing bucket with missing template bucket', async () => {
     const storage = new AWSStorage({
       AWS_S3_REGION: 'foo',
       AWS_S3_ACCESS_KEY_ID: 'bar',
       AWS_S3_SECRET_ACCESS_KEY: 'baz',
       mount: { url: 'mymount' },
     });
+    storage.client.storage = new Map();
+
+    await assert.rejects(() => storage.store(
+      'live', '/path', new Response('body', { status: 200 }),
+    ));
+  });
+  it('store 2 items to non existing bucket', async () => {
+    const storage = new AWSStorage({
+      AWS_S3_REGION: 'foo',
+      AWS_S3_ACCESS_KEY_ID: 'bar',
+      AWS_S3_SECRET_ACCESS_KEY: 'baz',
+      mount: { url: 'mymount' },
+    });
+    const memStorage = new Map();
+    memStorage.set('helix-content-bus-template', []);
+    storage.client.storage = memStorage;
+
     await assert.doesNotReject(() => storage.store(
-      '/path', new Response('body', { status: 200 }),
+      'live', '/path', new Response('body', {
+        status: 200,
+        headers: { 'last-modified': 'Tue, 20 Apr 2021 23:51:03 GMT' },
+      }),
+    ));
+    await assert.doesNotReject(() => storage.store(
+      'live', '/path2', new Response('body', { status: 200 }),
+    ));
+  });
+
+  it('store item to existing bucket', async () => {
+    const storage = new AWSStorage({
+      AWS_S3_REGION: 'foo',
+      AWS_S3_ACCESS_KEY_ID: 'bar',
+      AWS_S3_SECRET_ACCESS_KEY: 'baz',
+      mount: { url: 'mymount2' },
+    });
+    const memStorage = new Map();
+    memStorage.set('h3c4fbb7d701e130329d716baabc51d95ef3a9fb6bbc2df1f469caf2150fd', []);
+    storage.client.storage = memStorage;
+
+    await assert.doesNotReject(() => storage.store(
+      'live', '/path', new Response('body', { status: 200 }),
     ));
   });
 });